Add files via upload
We have a string consisting of an equal number of open and closed parentheses. At each step, we can change the place of two adjacent parentheses. We want to reach a state where the string containing the parentheses is ordered. A string is ordered if it is 1(2)
be S1(S2). S1 and S2 are ordered strings. Also, the null string is considered to be ordered. What is the minimum number of steps to do this?